Unifying into single generate script
