+++ generate-workflow.py
@@ -5,6 +5,7 @@
 # jobs with similar content and as far as we know, the workflow language has
 # essentially no support for code reuse. :(
 
+import os
 import textwrap
 from typing import List, NamedTuple, Optional
 
@@ -195,11 +196,11 @@ benchmarks = [
 ]
 
 HEADER = '''\
-# This file is generated by generate-workflow.py. To update this file, update
-# generate-workflow.py instead and re-run it. Some things in this file are
-# explained by comments in generate-workflow.py.
+# This file is generated by generate-exhaustive-workflow.py. To update this file, update
+# generate-exhaustive-workflow.py instead and re-run it. Some things in this file are
+# explained by comments in generate-exhaustive-workflow.py.
 
-name: 3C benchmark tests
+name: Exhaustive testing and Performance Stats 
 
 on:
   # Run every day at 3:00 AM EDT (i.e., 7 am UTC)
@@ -323,92 +324,170 @@ class Step(NamedTuple):
         return textwrap.indent(part1 + part2, 6 * ' ')
 
 
+class ActionStep:
+    def __init__(self, act_name, use_act, act_args):
+        self.act_name = act_name
+        self.use_act = use_act
+        self.act_args = act_args
+
+    def __str__(self):
+        part1 = textwrap.dedent(f'''\
+        - name: {self.act_name}
+          uses: {self.use_act}
+          with: 
+        ''')
+        all_args = ''
+        for arg_key in self.act_args:
+            all_args += arg_key + ": " + f'''{self.act_args[arg_key]}''' + '\n'
+        part2 = textwrap.indent(all_args, 4 * ' ')
+        return textwrap.indent(part1 + part2, 6 * ' ')
+
+
 def ensure_trailing_newline(s: str):
     return s + '\n' if s != '' and not s.endswith('\n') else s
 
 
-with open('.github/workflows/main.yml', 'w') as out:
-    out.write(HEADER)
-    for binfo in benchmarks:
-        for alltypes in (False, True):
-            at_dir = ('${{env.benchmark_conv_dir}}/' +
-                      ('alltypes' if alltypes else 'no-alltypes'))
-            at_job = 'alltypes' if alltypes else 'no_alltypes'
-            at_job_friendly = '-alltypes' if alltypes else 'no -alltypes'
-            convert_extra = (ensure_trailing_newline(binfo.convert_extra)
-                             if binfo.convert_extra is not None else '')
-            build_converted_cmd = binfo.build_converted_cmd.rstrip('\n')
-            # Python argparse thinks `--extra-3c-arg -alltypes` is two options
-            # rather than an option with an argument.
-            at_flag = '--extra-3c-arg=-alltypes \\\n' if alltypes else ''
-            at_ignore_step = ' (ignore failure)' if alltypes else ''
-            at_ignore_code = ' || true' if alltypes else ''
+def create_benchmark(out, binfo, alltypes, generate_stats=False, extra_args=[]):
+    # Python argparse thinks `--extra-3c-arg -alltypes` is two options
+    # rather than an option with an argument.
+    at_flag = '--extra-3c-arg=-alltypes \\\n' if alltypes else ''
 
-            out.write(f'''\
+    job_suffix = ''
+    for earg in extra_args:
+        if not job_suffix:
+            job_suffix = "_"
+        at_flag += '--extra-3c-arg=' + earg + ' \\\n'
+        job_suffix += earg.replace('-', '_')
 
-  test_{binfo.name}_{at_job}:
-    name: Test {binfo.friendly_name} ({at_job_friendly})
+    at_dir = ('${{env.benchmark_conv_dir}}/' +
+              ('alltypes' if alltypes else 'no-alltypes') + job_suffix)
+    at_job = 'alltypes' if alltypes else 'no_alltypes'
+    at_job_friendly = '-alltypes' if alltypes else 'no -alltypes'
+    convert_extra = (ensure_trailing_newline(binfo.convert_extra)
+                     if binfo.convert_extra is not None else '')
+    build_converted_cmd = binfo.build_converted_cmd.rstrip('\n')
+    at_ignore_step = ' (ignore failure)' if alltypes else ''
+    at_ignore_code = ' || true' if alltypes else ''
+
+    out.write(f'''\
+  test_{binfo.name}_{at_job}{job_suffix}:
+    name: Test {binfo.friendly_name} ({at_job_friendly}{job_suffix})
     needs: build_3c
     runs-on: self-hosted
     steps:
 ''')
 
-            full_build_cmds = textwrap.dedent(f'''\
-            mkdir -p {at_dir}
-            cd {at_dir}
-            tar -xvzf ${{{{env.benchmark_tar_dir}}}}/{binfo.dir_name}.tar.gz
-            cd {binfo.dir_name}
-            ''') + ensure_trailing_newline(binfo.build_cmds)
+    full_build_cmds = textwrap.dedent(f'''\
+                mkdir -p {at_dir}
+                cd {at_dir}
+                tar -xvzf ${{{{env.benchmark_tar_dir}}}}/{binfo.dir_name}.tar.gz
+                cd {binfo.dir_name}
+                ''') + ensure_trailing_newline(binfo.build_cmds)
 
-            steps = [Step('Build ' + binfo.friendly_name, full_build_cmds)]
+    steps = [Step(
+                 'Build ' + binfo.friendly_name,
+                 full_build_cmds)]
 
-            components = binfo.components
-            if components is None:
-                components = [BenchmarkComponent(binfo.friendly_name)]
+    components = binfo.components
+    if components is None:
+        components = [BenchmarkComponent(binfo.friendly_name)]
 
-            for component in components:
-                component_dir = f'{at_dir}/{binfo.dir_name}'
-                if component.subdir is not None:
-                    component_dir += '/' + component.subdir
-                component_friendly_name = (component.friendly_name or
-                                           binfo.friendly_name)
+    for component in components:
+        component_dir = f'{at_dir}/{binfo.dir_name}'
+        if component.subdir is not None:
+            component_dir += '/' + component.subdir
+        component_friendly_name = (component.friendly_name or
+                                   binfo.friendly_name)
 
-                # yapf: disable
-                convert_flags = textwrap.indent(
-                    convert_extra +
-                    '--includeDir ${{env.include_dir}} \\\n' +
-                    '--prog_name ${{env.builddir}}/bin/3c \\\n' +
-                    at_flag +
-                    '--project_path .' +
-                    (f' \\\n--build_dir {component.build_dir}'
-                     if component.build_dir is not None else '') +
-                    '\n',
-                    2 * ' ')
-                # yapf: enable
-                steps.append(
-                    Step(
-                        'Convert ' + component_friendly_name,
-                        textwrap.dedent(f'''\
-                        cd {component_dir}
-                        ${{{{env.port_tools}}}}/convert_project.py \\
-                        ''') + convert_flags))
+        # yapf: disable
+        convert_flags = textwrap.indent(
+            convert_extra +
+            '--includeDir ${{env.include_dir}} \\\n' +
+            '--prog_name ${{env.builddir}}/bin/3c \\\n' +
+            at_flag +
+            '--project_path .' +
+            (f' \\\n--build_dir {component.build_dir}'
+             if component.build_dir is not None else '') +
+            '\n',
+            2 * ' ')
+        # yapf: enable
+        steps.append(
+            Step(
+                'Convert ' + component_friendly_name,
+                textwrap.dedent(f'''\
+                            cd {component_dir}
+                            ${{{{env.port_tools}}}}/convert_project.py \\
+                            ''') + convert_flags))
 
-                steps.append(
-                    Step(
-                        'Build converted ' + component_friendly_name +
-                        at_ignore_step,
-                        # convert_project.py sets -output-dir=out.checked as
-                        # standard.
-                        textwrap.dedent(f'''\
-                        cd {component_dir}
-                        cp -r out.checked/* .
-                        rm -r out.checked
-                        ''') +
-                        #
-                        (f'cd {component.build_dir}\n'
-                         if component.build_dir is not None else '') +
-                        f'{build_converted_cmd}{at_ignore_code}\n'))
+        if generate_stats:
+            perf_dir_name = "3c_performance_stats/"
+            steps.append(
+                Step(
+                    'Copy 3c stats of ' + component_friendly_name,
+                    textwrap.dedent(f'''\
+                                cd {component_dir}
+                                mkdir {perf_dir_name}
+                                cp *.json {perf_dir_name}
+                                ''')))
+            perf_artifact_name = component_friendly_name + job_suffix + ('_alltypes' if alltypes else '_noalltypes')
+            perf_dir = os.path.join(component_dir, perf_dir_name)
+            steps.append(
+                ActionStep(
+                    'Upload 3c stats of ' + component_friendly_name,
+                    'actions/upload-artifact@v2',
+                    {'name': perf_artifact_name, 'path': perf_dir, 'retention-days': 5}))
 
-            # We want blank lines between steps but not after the last step of
-            # the last benchmark.
-            out.write('\n'.join(str(s) for s in steps))
+        steps.append(
+            Step(
+                'Build converted ' + component_friendly_name +
+                at_ignore_step,
+                # convert_project.py sets -output-dir=out.checked as
+                # standard.
+                textwrap.dedent(f'''\
+                            cd {component_dir}
+                            cp -r out.checked/* .
+                            rm -r out.checked
+                            ''') +
+                #
+                (f'cd {component.build_dir}\n'
+                 if component.build_dir is not None else '') +
+                f'{build_converted_cmd}{at_ignore_code}\n'))
+
+    # We want blank lines between steps but not after the last step of
+    # the last benchmark.
+    out.write('\n'.join(str(s) for s in steps) + '\n')
+
+
+regular_run_configurations = {
+    False: [],
+    True: []
+}
+
+exhaustive_run_configurations = {
+    False: ['-only-g-sol', '-only-l-sol'],
+    True: []
+}
+
+'''
+Format:
+  {
+    workflow_file_path : (<run_config>, [True|False] <- flag to enable stats generation)
+  }
+'''
+workflow_file_configs = {
+    '.github/workflows/main.yml': (regular_run_configurations, False),
+    '.github/workflows/exhaustive.yml': (exhaustive_run_configurations, True)
+}
+
+for workflow_file in workflow_file_configs:
+    run_configurations = workflow_file_configs[workflow_file][0]
+    generate_stats = workflow_file_configs[workflow_file][1]
+    with open(workflow_file, 'w') as out:
+        out.write(HEADER)
+        for binfo in benchmarks:
+            for alltypes in run_configurations:
+                create_benchmark(out, binfo, alltypes, generate_stats)
+                extra_args = run_configurations[alltypes]
+                if len(extra_args) > 0:
+                    for earg in extra_args:
+                        create_benchmark(out, binfo, alltypes, generate_stats, [earg])
